+++ black_mouth_kinematics/CMakeLists.txt
@@ -21,7 +21,6 @@ find_package(Eigen3 REQUIRED)
 find_package(rclcpp REQUIRED)
 find_package(tf2_eigen REQUIRED)
 find_package(std_msgs REQUIRED)
-find_package(sensor_msgs REQUIRED)
 find_package(geometry_msgs REQUIRED)
 find_package(trajectory_msgs REQUIRED)
 find_package(rosidl_default_generators REQUIRED)
@@ -50,16 +49,12 @@ ament_target_dependencies(inv_kinematics_node rclcpp std_msgs trajectory_msgs)
 add_executable(inv_kinematics_trajectory test/publish_ik_trajectory.cpp)
 ament_target_dependencies(inv_kinematics_trajectory rclcpp)
 
-add_executable(joy_body_ik src/JoyBodyIK.cpp)
-ament_target_dependencies(joy_body_ik rclcpp sensor_msgs)
-
 rosidl_get_typesupport_target(cpp_typesupport_target "${PROJECT_NAME}" "rosidl_typesupport_cpp")
 
 target_link_libraries(inv_kinematics_server "${cpp_typesupport_target}")
 target_link_libraries(inv_kinematics_client "${cpp_typesupport_target}")
 target_link_libraries(inv_kinematics_node "${cpp_typesupport_target}")
 target_link_libraries(inv_kinematics_trajectory "${cpp_typesupport_target}")
-target_link_libraries(joy_body_ik "${cpp_typesupport_target}")
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
@@ -86,7 +81,6 @@ install(
   inv_kinematics_client
   inv_kinematics_node
   inv_kinematics_trajectory
-  joy_body_ik
   DESTINATION 
   lib/${PROJECT_NAME}
 )
