+++ black_mouth_kinematics/package.xml
@@ -18,12 +18,10 @@
 
   <depend>rclcpp</depend>
   <depend>std_msgs</depend>
-  <depend>sensor_msgs</depend>
   <depend>geometry_msgs</depend>
   <depend>trajectory_msgs</depend>
   <depend>tf2_eigen</depend>
   <depend>eigen</depend>
-  <depend>joy</depend> 
 
   <member_of_group>rosidl_interface_packages</member_of_group>
