+++ black_mouth_teleop/msg/States.msg
@@ -1,0 +1,3 @@
+uint8 STATE0=0
+uint8 STATE1=0
+uint8 STATE2=0
