+++ black_mouth_teleop/package.xml
@@ -3,14 +3,26 @@
 <package format="3">
   <name>black_mouth_teleop</name>
   <version>0.0.0</version>
-  <description>TODO: Package description</description>
+  <description>Black Mouth Teleop package</description>
   <maintainer email="felipe18mohr@gmail.com">Felipe Mohr</maintainer>
   <license>TODO: License declaration</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
+  <build_depend>rosidl_default_generators</build_depend>
+
+  <exec_depend>rosidl_default_runtime</exec_depend>
+
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
+
+  <depend>rclcpp</depend>
+  <depend>std_msgs</depend>
+  <depend>sensor_msgs</depend>
+  <depend>black_mouth_kinematics</depend>
+  <depend>joy</depend> 
+
+  <member_of_group>rosidl_interface_packages</member_of_group>
 
   <export>
     <build_type>ament_cmake</build_type>
